Fix: 부동 소숫점 이슈 BigDecimal로 해결

+++ src/main/java/com/doubledeltas/minecollector/config/chapter/ScoringChapter.java
@@ -3,17 +3,18 @@ package com.doubledeltas.minecollector.config.chapter;
 import lombok.Data;
 import org.bukkit.advancement.AdvancementDisplayType;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 public class ScoringChapter {
-    private boolean collectionEnabled;
-    private float   collectionScore;
+    private boolean     collectionEnabled;
+    private BigDecimal  collectionScore;
 
-    private boolean stackEnabled;
-    private int     stackMultiple;
-    private float   stackScore;
+    private boolean     stackEnabled;
+    private int         stackMultiple;
+    private float       stackScore;
 
-    private boolean advancementEnabled;
-    private Map<AdvancementDisplayType, Float> advancementScores;
+    private boolean     advancementEnabled;
+    private Map<AdvancementDisplayType, BigDecimal> advancementScores;
 }
